backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
weasyprint==62.3
matplotlib==3.9.0
httpx==0.27.0
pydantic==2.9.0
slowapi==0.1.9
# v2.0 — multi-format parsers
openpyxl==3.1.5
xlrd==2.0.1
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
pillow-heif==0.18.0
